Restoring original development section on index page.
First, this sections target audience is not users, but
technical people, so the aim is not to make it
comprehensible at first read for 'new readers', but
the way people expect it that work with the stuff
(and there, the sorting is based on source package name).

Second, the section is named development on purpose,
we are reusing the same scheme that debian.org does.

Third, no problem in using something different or new,
but 'components' isn't fitting well either. That section
contains the technical stuff regarding development,
users interested just in how things work should not look
there for information, but in the manual. This section
really is about development.

And last but not least.. it would be much nicer to talk
about such things on the mailinglist first, before
you (Jonathan) just edit someone else index page entirely.